MSS-118 Relative paths now get ./ prefix
- Changed DicomSource to use the `./` relative prefix when returning relative paths (to the ArchiveRoot)
- Added tests for zip file (using BadMedicine.Dicom)
