+++ Rdmp.Dicom.Tests/Rdmp.Dicom.Tests.csproj
@@ -36,6 +36,7 @@
     <ProjectReference Include="..\Rdmp.Dicom\Rdmp.Dicom.csproj" />
   </ItemGroup>
   <ItemGroup>
+    <PackageReference Include="HIC.BadMedicine.Dicom" Version="0.0.4" />
     <PackageReference Include="HIC.RDMP.Plugin.Test" Version="4.0.1" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.2.0" />
     <PackageReference Include="NUnit" Version="3.12.0" />
